requirements.txt
fastapi
uvicorn
python-multipart
Tortoise-orm